Extract formatCount to a lib
